# Improve OptimizedSolver