--- a/pentomino-solver/Cargo.toml
+++ b/pentomino-solver/Cargo.toml
@@ -6,5 +6,6 @@ edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
+derive_more = "0.99"
 num-traits = "0.2"
 num-derive = "0.4"