--- a/pentomino-solver/src/bitboard.rs
+++ b/pentomino-solver/src/bitboard.rs
@@ -1,39 +1,25 @@
-use std::ops::{BitAnd, BitOr};
+use derive_more::{BitAnd, BitAndAssign, BitOr, BitOrAssign, From, Into};
 
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Default)]
+#[derive(
+    Debug,
+    Clone,
+    Copy,
+    PartialEq,
+    Eq,
+    Default,
+    // From/Into, and bit manipulations by derive_more
+    From,
+    Into,
+    BitAnd,
+    BitAndAssign,
+    BitOr,
+    BitOrAssign,
+)]
 pub struct Bitboard(u64);
 
 impl Bitboard {
     pub fn is_empty(&self) -> bool {
         self.0 == 0
-    }
-}
-
-impl From<u64> for Bitboard {
-    fn from(value: u64) -> Self {
-        Self(value)
-    }
-}
-
-impl From<Bitboard> for u64 {
-    fn from(b: Bitboard) -> Self {
-        b.0
-    }
-}
-
-impl BitAnd for Bitboard {
-    type Output = Self;
-
-    fn bitand(self, rhs: Self) -> Self::Output {
-        Self(self.0 & rhs.0)
-    }
-}
-
-impl BitOr for Bitboard {
-    type Output = Self;
-
-    fn bitor(self, rhs: Self) -> Self::Output {
-        Self(self.0 | rhs.0)
     }
 }