--- a/pentomino-solver/src/pieces.rs
+++ b/pentomino-solver/src/pieces.rs
@@ -1,9 +1,9 @@
-use num_derive::FromPrimitive;
+use num_derive::{FromPrimitive, ToPrimitive};
 use std::fmt::Write;
 
 pub const NUM_PIECES: usize = 12;
 
-#[derive(Debug, Copy, Clone, PartialEq, Eq, Hash, FromPrimitive)]
+#[derive(Debug, Copy, Clone, PartialEq, Eq, Hash, FromPrimitive, ToPrimitive)]
 pub enum Piece {
     O = 0,
     P,