Benchmarking by MacBook Pro 14-inch, 2021 Apple M1 Pro 32 GB

before:
```
test bench_8x8_2x2_default          ... bench: 309,469,066 ns/iter (+/- 4,766,181)
test bench_8x8_2x2_default_unique   ... bench: 310,233,200 ns/iter (+/- 5,945,207)
test bench_8x8_2x2_optimized        ... bench: 301,117,020 ns/iter (+/- 2,549,747)
test bench_8x8_2x2_optimized_unique ... bench: 303,714,658 ns/iter (+/- 1,891,695)
test bench_8x8_2x2_simple           ... bench: 313,255,645 ns/iter (+/- 1,770,687)
```

after:
```
test bench_3x20_optimized           ... bench:     388,075 ns/iter (+/- 2,925)
test bench_3x20_optimized_unique    ... bench:     382,584 ns/iter (+/- 10,356)
test bench_4x15_optimized           ... bench:   9,387,291 ns/iter (+/- 126,266)
test bench_4x15_optimized_unique    ... bench:   8,437,287 ns/iter (+/- 163,549)
test bench_5x12_optimized           ... bench:  44,823,054 ns/iter (+/- 488,594)
test bench_5x12_optimized_unique    ... bench:  40,230,583 ns/iter (+/- 1,012,642)
test bench_6x10_optimized           ... bench: 108,035,254 ns/iter (+/- 1,079,854)
test bench_6x10_optimized_unique    ... bench:  98,153,708 ns/iter (+/- 1,064,474)
test bench_8x8_2x2_default          ... bench: 306,786,462 ns/iter (+/- 1,432,446)
test bench_8x8_2x2_default_unique   ... bench: 309,572,341 ns/iter (+/- 3,754,899)
test bench_8x8_2x2_optimized        ... bench:  22,005,954 ns/iter (+/- 225,645)
test bench_8x8_2x2_optimized_unique ... bench:  21,487,987 ns/iter (+/- 333,553)
test bench_8x8_2x2_simple           ... bench: 323,421,933 ns/iter (+/- 3,125,835)
```